+++ core/src/routes/(editor)/editor/code.svelte
@@ -4,5 +4,5 @@
 
 <div class="w-full">
     <pre>{JSON.stringify($globalComponentCollectionStore, null, 2)}</pre>
-
+    
 </div>