+++ core/src/routes/uicomponents/action-hover-button.svelte
@@ -55,7 +55,7 @@
      */
     let buttonPassiveBackgroundColor;
 
-    if(globalThemeStore){
+    if(globalThemeStore != undefined && $globalThemeStore != undefined){
         buttonActiveIconColor = $globalThemeStore.button.active.iconColor;
         buttonActiveBackgroundColor = $globalThemeStore.button.active.backgroundColor;
         buttonPassiveIconColor = $globalThemeStore.button.passive.iconColor;